feat: Remove TV show-specific use cases and group them into one use case
