--- domain/src/main/java/com/london/domain/usecase/GetTvShowDetails.kt
@@ -1,11 +1,0 @@
-package com.london.domain.usecase
-
-import com.london.domain.repository.TvShowRepository
-import javax.inject.Inject
-
-class GetTvShowDetails @Inject constructor(
-    private val tvShowRepository: TvShowRepository
-) {
-    suspend operator fun invoke(tvShowId: Int) =
-        tvShowRepository.getTvShowDetailsById(tvShowId)
-}
