--- domain/src/main/java/com/london/domain/usecase/GetTvShowVideoProvider.kt
@@ -1,10 +1,0 @@
-package com.london.domain.usecase
-
-import com.london.domain.repository.TvShowVideoProviderRepository
-import javax.inject.Inject
-
-class GetTvShowVideoProvider @Inject constructor(
-    private val repository: TvShowVideoProviderRepository
-) {
-    suspend fun invoke(tvShowId: Int) = repository.getTvShowVideos(tvShowId)
-}
