--- domain/src/main/java/com/london/domain/usecase/GetTvShowsByCategoryId.kt
@@ -1,15 +1,0 @@
-package com.london.domain.usecase
-
-import com.london.domain.repository.discover.DiscoverRepository
-import javax.inject.Inject
-
-class GetTvShowsByCategoryId @Inject constructor(
-    private val repository: DiscoverRepository
-) {
-    suspend fun invoke(
-        categoryId: Int, pageNumber: Int
-    ) = repository.getTvShowsByCategory(
-        categoryId = categoryId,
-        pageNumber = pageNumber
-    )
-}
