+++ domain/src/main/java/com/london/domain/usecase/details/tvshow/ManageTvShowDetailsUseCase.kt
@@ -1,0 +1,49 @@
+package com.london.domain.usecase.details.tvshow
+
+import com.london.domain.entity.PagedFetchResponse
+import com.london.domain.entity.Trending
+import com.london.domain.repository.PopularRepository
+import com.london.domain.repository.SearchRepository
+import com.london.domain.repository.TrendingRepository
+import com.london.domain.repository.TvShowRepository
+import com.london.domain.repository.TvShowVideoProviderRepository
+import com.london.domain.repository.discover.DiscoverRepository
+import javax.inject.Inject
+
+class ManageTvShowDetailsUseCase @Inject constructor(
+    private val tvShowRepository: TvShowRepository,
+    private val popularRepository: PopularRepository,
+    private val trendingRepository: TrendingRepository,
+    private val searchRepository: SearchRepository,
+    private val discoverRepository: DiscoverRepository,
+    private val tvShowVideoProviderRepository: TvShowVideoProviderRepository
+) {
+    suspend fun getTvShowDetails(tvShowId: Int) = tvShowRepository.getTvShowDetailsById(tvShowId)
+
+    suspend fun getPopularTvShows(limit: Int = LIMIT) =
+        popularRepository.getPopularTvShows().take(limit)
+
+    suspend fun getTrendingTvShows(page: Int): PagedFetchResponse<Trending> =
+        trendingRepository.getTrendingTvShows(page = page)
+
+    suspend fun getTvShowList(name: String, pageNumber: Int) =
+        searchRepository.searchForTvShows(
+            name = name,
+            pageNumber = pageNumber
+        )
+
+    suspend fun getTvShowsByCategory(
+        categoryId: Int, pageNumber: Int
+    ) = discoverRepository.getTvShowsByCategory(
+        categoryId = categoryId,
+        pageNumber = pageNumber
+    )
+
+    suspend fun getTvShowVideoProvider(tvShowId: Int) =
+        tvShowVideoProviderRepository.getTvShowVideos(tvShowId)
+
+
+    companion object {
+        const val LIMIT = 5
+    }
+}
