+++ presentation/src/main/java/com/london/presentation/feature/details/tvshow/episodedetails/EpisodeDetailsViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.SavedStateHandle
 import com.london.domain.usecase.GetEpisodeByTvShowId
 import com.london.domain.usecase.GetEpisodeVideoProviderUseCase
 import com.london.domain.usecase.GetImagesById
-import com.london.domain.usecase.GetTvShowDetails
+import com.london.domain.usecase.details.tvshow.ManageTvShowDetailsUseCase
 import com.london.presentation.navigation.Screen
 import com.london.presentation.navigation.getArgs
 import com.london.presentation.shared.base.BaseViewModel
@@ -15,7 +15,7 @@ import javax.inject.Inject
 class EpisodeDetailsViewModel @Inject constructor(
     private val getTvShowImages: GetImagesById,
     private val getEpisodeByTvShowIdUseCase: GetEpisodeByTvShowId,
-    private val getTvShowDetails: GetTvShowDetails,
+    private val manageTvShowDetailsUseCase: ManageTvShowDetailsUseCase,
     private val getVideoProvider: GetEpisodeVideoProviderUseCase,
     savedStateHandle: SavedStateHandle,
 ) : BaseViewModel<EpisodeDetailsUiState, EpisodeDetailsEffect>(EpisodeDetailsUiState()),
@@ -38,7 +38,7 @@ class EpisodeDetailsViewModel @Inject constructor(
                     tvShowId, seasonNumber, episodeNumber
                 )
                 val images = getTvShowImages(tvShowId)
-                val tvShowDetails = getTvShowDetails(tvShowId)
+                val tvShowDetails = manageTvShowDetailsUseCase.getTvShowDetails(tvShowId)
 
 
                 Triple(episode, images, tvShowDetails)
