+++ presentation/src/main/java/com/london/presentation/feature/home/trending/tvshow/TrendingTvShowsViewModel.kt
@@ -1,6 +1,6 @@
 package com.london.presentation.feature.home.trending.tvshow
 
-import com.london.domain.usecase.GetTrendingTvShowsUseCase
+import com.london.domain.usecase.details.tvshow.ManageTvShowDetailsUseCase
 import com.london.presentation.shared.base.BaseViewModel
 import com.london.presentation.shared.base.createPagingSourceFlow
 import com.london.presentation.utils.TvShowGenre
@@ -8,7 +8,9 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class TrendingTvShowsViewModel @Inject constructor(private val getTrendingTvShows: GetTrendingTvShowsUseCase) :
+class TrendingTvShowsViewModel @Inject constructor(
+    private val manageTvShowDetailsUseCase: ManageTvShowDetailsUseCase,
+) :
     BaseViewModel<TrendingTvShowsUiState, TrendingTvShowsEffect>(TrendingTvShowsUiState()),
     TrendingTvShowsContract {
 
@@ -20,12 +22,13 @@ class TrendingTvShowsViewModel @Inject constructor(private val getTrendingTvShow
         tryToExecute(
             block = {
                 val tvShowsFlow = createPagingSourceFlow(query = "") { _, pageNumber ->
-                    val tvShows = getTrendingTvShows.invoke(page = pageNumber)
-                    val filteredItems = if (state.value.selectedGenreId != null && state.value.selectedGenreId != -1) {
-                        tvShows.items.filter { it.genreIds.contains(state.value.selectedGenreId) }
-                    } else {
-                        tvShows.items
-                    }
+                    val tvShows = manageTvShowDetailsUseCase.getTrendingTvShows(page = pageNumber)
+                    val filteredItems =
+                        if (state.value.selectedGenreId != null && state.value.selectedGenreId != -1) {
+                            tvShows.items.filter { it.genreIds.contains(state.value.selectedGenreId) }
+                        } else {
+                            tvShows.items
+                        }
                     tvShows.copy(items = filteredItems)
                 }
                 tvShowsFlow
